+++ common/rmi/src/main/java/is/codion/common/rmi/server/package-info.java
@@ -32,9 +32,9 @@
  * <li>{@link is.codion.common.rmi.server.ServerConfiguration#IDLE_CONNECTION_TIMEOUT}
  * <li>{@link is.codion.common.rmi.server.ServerConfiguration#AUXILIARY_SERVER_FACTORY_CLASS_NAMES}
  * <li>{@link is.codion.common.rmi.server.ServerConfiguration#OBJECT_INPUT_FILTER_FACTORY_CLASS_NAME}
+ * <li>{@link is.codion.common.rmi.server.SerializationFilterFactory#SERIALIZATION_FILTER_PATTERNS}
  * <li>{@link is.codion.common.rmi.server.SerializationFilterFactory#SERIALIZATION_FILTER_PATTERN_FILE}
- * <li>{@link is.codion.common.rmi.server.SerializationFilterFactory#SERIALIZATION_FILTER_DRYRUN_OUTPUT}
- * <li>{@link is.codion.common.rmi.server.SerializationFilterFactory#SERIALIZATION_FILTER_DRYRUN}
+ * <li>{@link is.codion.common.rmi.server.SerializationFilterFactory#SERIALIZATION_FILTER_DRYRUN_FILE}
  * </ul>
  */
 @org.jspecify.annotations.NullMarked