+++ changelog.md
@@ -14,6 +14,7 @@ Codion Change Log
 - MultiSelection.Indexes.set(Collection) removed.
 ### is.codion.common.rmi
 - SerializationWhitelist renamed SerializationFilter, whitelist implementation removed, now based on the built in pattern based filter.
+- SerializationFilterFactory, configuration properties renamed, some cleanup, docs improved.
 ### is.codion.swing.common.ui
 - ColumnConditionPanel.ComponentFactory, component() methods replaced with equal(), upper(), lower() and in().
 - LookAndFeelEnabler.enableLookAndFeel() now fallbacks to the defaultLookAndFeel, if the user preference one is not available.